+++ GamePatches/ObjectDumper.cs
@@ -1,0 +1,162 @@
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.IO;
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Text;
+using UnityEngine;
+
+namespace Recycle_N_Reclaim.GamePatches
+{
+    // from https://stackoverflow.com/a/10478008/6142494
+    // simple dumper for debug dumping. don't want to pull a json lib
+    public class ObjectDumper
+    {
+        private readonly List<int> _hashListOfFoundElements;
+        private readonly int _indentSize;
+        private readonly StringBuilder _stringBuilder;
+        private int _level;
+
+        private ObjectDumper(int indentSize)
+        {
+            _indentSize = indentSize;
+            _stringBuilder = new StringBuilder();
+            _hashListOfFoundElements = new List<int>();
+        }
+
+        public static string Dump(object element, int indentSize = 2)
+        {
+            var instance = new ObjectDumper(indentSize);
+            return instance.DumpElement(element);
+        }
+
+        private string GetTypeName(Type type)
+        {
+            return IsAnonymousType(type) ? "AnonymousType" : type.Name;
+        }
+
+        private string DumpElement(object element)
+        {
+            if (element is null or ValueType or string)
+            {
+                Write(FormatValue(element));
+            }
+            else
+            {
+                var objectType = element.GetType();
+                if (!typeof(IEnumerable).IsAssignableFrom(objectType))
+                {
+                    Write("{{{0}}}", GetTypeName(objectType));
+                    _hashListOfFoundElements.Add(element.GetHashCode());
+                    _level++;
+                }
+
+                if (element is IEnumerable enumerableElement)
+                {
+                    foreach (var item in enumerableElement)
+                        if (item is IEnumerable and not string)
+                        {
+                            _level++;
+                            DumpElement(item);
+                            _level--;
+                        }
+                        else
+                        {
+                            if (!AlreadyTouched(item))
+                                DumpElement(item);
+                            else
+                                Write("{{{0}}} <-- bidirectional reference found", GetTypeName(item.GetType()));
+                        }
+                }
+                else
+                {
+                    var members = element.GetType().GetMembers(BindingFlags.Public | BindingFlags.Instance);
+                    foreach (var memberInfo in members)
+                    {
+                        var fieldInfo = memberInfo as FieldInfo;
+                        var propertyInfo = memberInfo as PropertyInfo;
+
+                        if (fieldInfo == null && propertyInfo == null)
+                            continue;
+
+                        var type = fieldInfo != null ? fieldInfo.FieldType : propertyInfo.PropertyType;
+                        var value = fieldInfo != null
+                            ? fieldInfo.GetValue(element)
+                            : propertyInfo.GetValue(element, null);
+
+                        if (type.IsValueType || type == typeof(string))
+                        {
+                            Write("{0}: {1}", memberInfo.Name, FormatValue(value));
+                        }
+                        else
+                        {
+                            var isEnumerable = typeof(IEnumerable).IsAssignableFrom(type);
+                            Write("{0}: {1}", memberInfo.Name, isEnumerable ? "..." : "{ }");
+
+                            var alreadyTouched = !isEnumerable && AlreadyTouched(value);
+                            _level++;
+                            if (!alreadyTouched)
+                                DumpElement(value);
+                            else
+                                Write("{{{0}}} <-- bidirectional reference found", GetTypeName(value.GetType()));
+                            _level--;
+                        }
+                    }
+                }
+
+                if (!typeof(IEnumerable).IsAssignableFrom(objectType)) _level--;
+            }
+
+            return _stringBuilder.ToString();
+        }
+
+        private bool AlreadyTouched(object value)
+        {
+            if (value == null)
+                return false;
+
+            var hash = value.GetHashCode();
+            for (var i = 0; i < _hashListOfFoundElements.Count; i++)
+                if (_hashListOfFoundElements[i] == hash)
+                    return true;
+            return false;
+        }
+
+        private void Write(string value, params object[] args)
+        {
+            var space = new string(' ', _level * _indentSize);
+
+            if (args != null)
+                value = string.Format(value, args);
+
+            _stringBuilder.AppendLine(space + value);
+        }
+
+        private string FormatValue(object o)
+        {
+            return o switch
+            {
+                null => "null",
+                DateTime time => time.ToShortDateString(),
+                string => $"\"{o}\"",
+                '\0' => string.Empty,
+                ValueType => o.ToString(),
+                IEnumerable => "...",
+                _ => "{ }"
+            };
+        }
+
+        private static bool IsAnonymousType(Type type)
+        {
+            if (type == null)
+                throw new ArgumentNullException("type");
+
+            // HACK: The only way to detect anonymous types right now.
+            return Attribute.IsDefined(type, typeof(CompilerGeneratedAttribute), false)
+                   && type.IsGenericType && type.Name.Contains("AnonymousType")
+                   && (type.Name.StartsWith("<>") || type.Name.StartsWith("VB$"))
+                   && type.Attributes.HasFlag(TypeAttributes.NotPublic);
+        }
+    }
+}
