+++ ILRepack.targets
@@ -6,6 +6,7 @@
             <InputAssemblies Include="$(OutputPath)\ServerSync.dll" />
             <InputAssemblies Include="$(OutputPath)\YamlDotNet.dll" />
             <InputAssemblies Include="$(OutputPath)\Libs\JewelcraftingAPI.dll" />
+            <InputAssemblies Include="$(OutputPath)\Libs\AugaAPI.dll" />
             <!-- Only use the line below if you choose to use the DLL and not the classes. If you choose the DLL, be sure to reference and 
              tell it to copy to output directory-->
             <!-- <InputAssemblies Include="$(OutputPath)\StatusEffectManager.dll" /> -->
